pyproject.toml
[tool.poetry]
name = "lechefacil-backend"
version = "0.1.0"
description = "Minimal multi-tenant backend MVP for LecheFacil with FastAPI"
authors = ["LecheFacil Team <team@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
httpx = "^0.27.0"
SQLAlchemy = { version = "^2.0.29", extras = ["asyncio"] }
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
botocore = "^1.34.0"
boto3 = "^1.34.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
reportlab = "^4.0.0"
plotly = "^5.17.0"

[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.19.0"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
ruff = "^0.4.1"

[tool.ruff]
line-length = 100
exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "ASYNC"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
